# Fly/FlyAPI/Model/FinalTicket.cs
namespace FlyAPI.Model
{
    public class FinalTicket
    {

        

        public string FirstName { get; set; }
        public string LastName { get; set; }
        public int FlightNbr { get; set; }
        public double Price { get; set; }
    }
}