# FlyAlex/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "paths": {
    "/GetAvailableFlights": {
      "get": {
        "tags": [
          "Flights"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/GetPriceFromFlight/{flightId}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "application/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "text/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    },
    "/BookFlight/{passengerId}&{flightId}": {
      "put": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "passengerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/GetTotalSalePriceOfFlight/{flightId}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "application/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "text/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    },
    "/AverageTicketPrice/{destination}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "destination",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "application/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              },
              "text/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    },
    "/FinalTicket/{destination}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "destination",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FinalTicket"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FinalTicket"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FinalTicket"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Baggage": {
        "type": "object",
        "properties": {
          "baggageID": {
            "type": "integer",
            "format": "int32"
          },
          "baggageWeight": {
            "type": "number",
            "format": "float"
          },
          "baggageSizer": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "Booking": {
        "type": "object",
        "properties": {
          "bookingID": {
            "type": "integer",
            "format": "int32"
          },
          "salePrice": {
            "type": "number",
            "format": "double"
          },
          "passenger": {
            "$ref": "#/components/schemas/Passenger"
          },
          "flight": {
            "$ref": "#/components/schemas/Flight"
          }
        },
        "additionalProperties": false
      },
      "FinalTicket": {
        "type": "object",
        "properties": {
          "ticketId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "flightNbr": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "Flight": {
        "type": "object",
        "properties": {
          "flightId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "depart": {
            "type": "string",
            "nullable": true
          },
          "destination": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "flightNo": {
            "type": "integer",
            "format": "int32"
          },
          "occupation": {
            "type": "number",
            "format": "double"
          },
          "nbrSeat": {
            "type": "integer",
            "format": "int32"
          },
          "freeSeat": {
            "type": "integer",
            "format": "int32"
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Passenger": {
        "type": "object",
        "properties": {
          "personID": {
            "type": "integer",
            "format": "int32"
          },
          "birthday": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "baggages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Baggage"
            },
            "nullable": true
          },
          "flights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Flight"
            },
            "nullable": true
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}